Abstract run and fix cmd flags
